+++ src/pages/mypage/MyEvents.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { useQuery } from 'react-query';
 
 import { Button } from '@components/ui/button';
 import SpinnerBox from '@shared/SpinnerBox';
@@ -8,28 +7,13 @@ import ErrorBox from '@shared/ErrorBox';
 import Status from '@components/mypage/events/Status';
 import EventTable from '@components/mypage/events/EventTable';
 
-import { EventType } from '@/types/event';
-import useUser from '@hooks/useUser';
-import { getMyEvents } from '@services/eventService';
 import { StatusMenuType } from '@constants/eventStatusMenus';
+import useMyEvents from '@/hooks/useMyEvents';
 
 function MyEvents() {
-  const user = useUser();
   const [status, setStatus] = useState<StatusMenuType | null>(null);
 
-  const { data, isLoading, isError } = useQuery<EventType[]>(
-    ['myEvents', user?.uid],
-    async () => {
-      if (!user?.uid) {
-        return [];
-      }
-      return await getMyEvents(user.uid);
-    },
-    {
-      enabled: !!user && !!user.uid,
-      cacheTime: 1000 * 60 * 30,
-    },
-  );
+  const { data, isLoading, isError } = useMyEvents();
 
   const filteredData = status
     ? data?.filter(event => event.status === status) || []
